requirements.txt
alabaster==0.7.10
asn1crypto==0.22.0
aws-requests-auth==0.4.0
awsebcli==3.12.1
Babel==2.5.0
beautifulsoup4==4.6.0
blessed==1.14.2
boto3==1.4.7
botocore==1.7.43
cairocffi==0.8.0
cement==2.8.2
certifi==2017.7.27.1
cffi==1.11.0
chardet==3.0.4
colorama==0.3.7
cryptography==2.0.3
csscompressor==0.9.4
Django==1.11.3
django-appconf==1.0.2
django-background-tasks==1.1.11
django-bootstrap3==9.0.0
django-bootstrap4==0.0.3
django-cacheops==4.0
django-compat==1.0.14
django-compressor==2.2
django-crispy-forms==1.6.1
django-csvimport==2.11
django-debug-toolbar==1.8
django-elasticsearch-dsl==0.2.0
django-environ-0.4.4
django-export-action==0.1.1
django-filter==1.0.4
django-haystack==2.6.1
django-htmlmin==0.10.0
django-ipware==1.1.6
django-livereload-server==0.2.3
django-mailgun==0.9.1
django-redis==4.8.0
django-redis-cache==1.7.1
django-report-utils==0.3.15
django-simple-redis-admin==1.4.0
django-storages==1.6.5
djangorestframework==3.7.3
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
elasticsearch==5.4.0
elasticsearch-dsl==5.3.0
et-xmlfile==1.0.1
funcy==1.9
google-api-python-client==1.6.4
html5lib==0.999999999
httplib2==0.10.3
idna==2.6
imagesize==0.7.1
jdcal==1.3
Jinja2==2.9.6
jmespath==0.9.3
keen==0.5.1
Markdown==2.6.9
MarkupSafe==1.0
model-mommy==1.4.0
ndg-httpsclient==0.4.3
oauth2client==4.1.2
olefile==0.44
openpyxl==2.4.9
pathspec==0.5.0
Pillow==4.2.1
ply==3.10
psycopg2==2.7.3
pyasn1==0.3.5
pyasn1-modules==0.1.4
pycparser==2.18
pycryptodome==3.4.7
Pygments==2.2.0
pyOpenSSL==17.3.0
PyScreeze==0.1.13
pysolr==3.6.0
python-dateutil==2.6.1
PyTweening==1.0.3
pytz==2017.2
PyYAML==3.12
raven==6.3.0
rcssmin==1.0.6
redis==2.10.6
requests==2.9.1
rjsmin==1.0.12
rsa==3.4.2
s3transfer==0.1.11
selenium==3.6.0
semantic-version==2.5.0
six==1.11.0
slimit==0.8.1
snowballstemmer==1.2.1
sqlparse==0.2.3
tabulate==0.7.5
termcolor==1.1.0
tornado==4.5.2
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.44.0
